api/requirements.txt
Django==4.2.14
django-cors-headers==3.5.0
django-filter==24.2
djangorestframework==3.15.2
factory-boy==3.3.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-django==4.8.0
